+++ src/main/java/site/travellaboratory/be/application/review/ReviewReaderService.java
@@ -10,14 +10,14 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import site.travellaboratory.be.common.exception.BeApplicationException;
 import site.travellaboratory.be.common.exception.ErrorCodes;
-import site.travellaboratory.be.infrastructure.domains.review.ReviewRepository;
+import site.travellaboratory.be.infrastructure.domains.review.repository.ReviewRepository;
 import site.travellaboratory.be.infrastructure.domains.review.entity.ReviewJpaEntity;
 import site.travellaboratory.be.domain.review.enums.ReviewStatus;
 import site.travellaboratory.be.infrastructure.domains.user.UserRepository;
 import site.travellaboratory.be.infrastructure.domains.user.entity.User;
 import site.travellaboratory.be.infrastructure.domains.user.enums.UserStatus;
-import site.travellaboratory.be.infrastructure.domains.userlikereview.ReviewLikeRepository;
-import site.travellaboratory.be.infrastructure.domains.userlikereview.entity.ReviewLikeJpaEntity;
+import site.travellaboratory.be.infrastructure.domains.review.repository.ReviewLikeRepository;
+import site.travellaboratory.be.infrastructure.domains.review.entity.ReviewLikeJpaEntity;
 import site.travellaboratory.be.domain.review.enums.ReviewLikeStatus;
 import site.travellaboratory.be.presentation.review.dto.reader.ProfileReviewLocation;
 import site.travellaboratory.be.presentation.review.dto.reader.ProfileReviewPaginationResponse;
